rds/requirements.txt
redis
pika
python-socketio
requests
websocket-client
